Make queries mostly immutable
All query methods (such as withTransaction) now return a new query object.

Most fields on query objects are now immutable.

The only exception are Pattern objects, which can still be accessed and then modified. Reasoner currently depends on this behaviour and changing it would require a large refactor (and may impact the performance of reasoner).