Bug #11044 testRestart concurrency issue
This bug crops up only when scheduling and cancelling a task in quick
succession.

When cancelling thread runTask would be executed (the thread wakes up).
This causes it to try and set its state to RUNNING at the same time as
the stop method tries to update it to STOPPED. Whichever thread does so
later overwrites the others TaskStatus update.

This fix adds a 100ms minimum delay before any task can be executed.