Analytics query syntax upgrade: use commas (not semicolon) and use square brackets for arguments (#2783)
_Note: this PR is big, but not because of refactorings, nor is it because it contains multiple features. It does contain refactorings, but you can see that they are only changes in function names and they are actually required because we want to align the terminology with the new grammar (and with the rest of the Graql Java API which already has it). This PR is actually big because of feature we are implementing (new analytics syntax) means that we need to change the way Graql Compute queries are being parsed. And given that we are implementing a new way of parsing the syntax, we have the opportunity to provide better error messages to our users and therefore this PR contains that too (rightly so)._

# Why is this PR needed?

We want to implement a new feature in Graql syntax, and that is the ability to provide (a set of) arguments to a given graql statement. Such as:
`define person sub entity, has email [unique, max=3];`
This goal requires us to first fix analytics so that the `graql compute` queries:
1. use commas (`,`) properly instead of (`;`), and 
2. use `[` and `]` to provide a list of arguments.

As a result, analytics queries will look like: 
`compute cluster in [person, marriage], using connected-component, where [start = "V123", members = true];`

However, the above is change is not possible in `graql compute` queries unless we change the grammar to be more flexible (composing it of generic patterns) rather than hard-coded patterns for each type of `graql compute` queries. The generic pattern for `graql compute` can now be simply described as:

Compute query -> `compute <method> <conditions>? ;`

The `Graql.g4` file contains a thorough description of the new grammar and should guide you to understand the whole implementation.

# What does the PR do?

Thus, this PR contains changes in:
1. Graql grammar definition (`Graql.g4`)
2. `GraqlConstructor` (renamed from `GraqlVisitor`) and `GraqlComputeConstructor` to implement the new way of parsing our `graql compute` queries.
3. New exceptions in `GraqlQueryException` along with the new error messages in `ErrorMessage`.
4. New implementations of `toString()` method for `graql compute` queries to also match the new syntax. The `toString()` method is composed of the same generic patterns: `method` and `conditions`, each of which are implemented uniquely in every `graql compute` query class.
5. Changes in function names in various `graql compute` query classes, to use the same terminology as the grammar file.

# Does it break backwards compatibility?

Yes, it does. Old `graql compute` syntax will not work with this new change.

# List of future improvements not on this PR

Nothing is known so far.