match-insert queries executes inserts before returning streams of answers (#2923)
# Why is this PR needed?

At the moment, `match-insert` queries does not perform the data insertion when you call the `.stream()`, because the stream hasn't been executed. Although this is programmatically correct, it is semantically wrong with regards to the expected behaviour of the query. The expected behaviour of `match-insert` `.stream()` is that the data insertion has been performed, but the results are yet to be streamed.

# What does the PR do?

We `.collect()` the answers of the current `match-insert` stream into a list so that the insertion is performed, but then we return the `.stream()` of the answer list again.

# Does it break backwards compatibility?

Nope.

# List of future improvements not on this PR
None.