Changed shell output test to be order agnostic (#2960)
# Why is this PR needed?

An aggregate group query test in GraqlShellIT tries to validate the output of a map, in which the order is not guaranteed. In the future, we should make the output order to be guaranteed, this is possible even with a map, and there is already a TP ticket for it. But for now, we can just fix the test so that it doesn't randomly fail.

# What does the PR do?

Replace the output "matchers" to check for any occurrence of map entry.