Improve the m2 cache handling from using CI cache to workflow. Fix failing test caused by an invalid CI cache (#2984)
# Why is this PR needed?
A test is failing due to bad handling of dependencies where some Grakn dependencies with snapshot versions (ie., `grakn-core x.x.x-SNAPSHOT) got incorrectly cached into the CircleCI cache.

# What does the PR do?
Switch the caching approach where we combine the cache together with the build, in the CircleCI workspace. This approach is simpler as well since we now only use CircleCI workspace instead of CircleCI caching AND workspace. This way, we don't have to deal with issues such as when to invalidate the cache based on checksum.

Bonus:
1. Set the deduplicator daemon log level to trace from debug
2. Removed an unused function in the DistributionContext

# Does it break backwards compatibility?
No

# List of future improvements not on this PR
N/A