Grouped and merged classes related to Sessions and Transactions class (#4600)
# Why is this PR needed?

We're in the process of redesigning Grakn architecture, which includes simplifying and improving how Sessions and Transactions are work (issue #3994). Right now, the implementation of Sessions and Transactions are too complex and unorganised. 

# What does the PR do?

The first to solve the problem above, we needed to: 
1. group all the related classes together in one package, so we can scope the problematic classes
2. rename the classes with consistent and meaningful names so we can quickly understand how they relate to each other
3. merge the classes that are too granularly modularised, so we can reduce the complexity

The result of the above changes is that we have all [implementation] classes related to Sessions and Transactions located in `grakn.core.server.session.*`, renamed and minimised.