Refactor Graql query language library - Part 2 (#4667)
# Why is this PR needed?

This PR is a continuation of PR #4646, and incremental work for issue #4456 where you can find the full description of the goal. In summary, we are refactoring Graql to extract the Graql language library from Grakn's implementation of Graql, and we start with simplifying the Graql codebase first.

# What does the PR do?

This PR is an intermediate step in simplifying the architecture of Graql Pattern Property classes (`grakn.core.graql.query.pattern.property.*`). We first simplify the unnecessary modularisation to make it easier to see the behaviour of the code.

- Make `IsAbstractProperty`, `RelatesProperty` and `RelationshipProperty` implement `NamedProperty`, so then we can:
- Merge `VarPropertyInternal` and `NamedProperty` interfaces
- Removed the usage `UniqueVarProperty`, in which `VarProperties` are down-casted to, and:
- Made sure all classes implemented `isUnique()` method
- Finally, we merged `VarProperty` (inteface), `VarPropertyInternal` (interface) and `AbstractVarProperty` (class), into just one abstract class: `VarProperty`

The result of the above change is depicted in the images below.

> Note: green lines indicate interface implementations, blue lines indicate class extensions

**Before**

![before](https://user-images.githubusercontent.com/567679/49187133-6f9f8200-f35e-11e8-877e-75b637ea2235.png)

**After**

![after](https://user-images.githubusercontent.com/567679/49187138-729a7280-f35e-11e8-8845-c8f388246e30.png)

# Does it break backwards compatibility?

This PR does not change any Graql behaviours.