Enable Bazel Remote Build Execution To Speed Up Tests (#4772)
# How To Use The Remote Build Executor (RBE)
## Enable RBE For Your Grakn Fork

RBE isn't enabled unless you add the appropriate credential in your fork. This is how you can do so:
1. Add your Grakn fork to CircleCI: https://circleci.com/add-projects/gh/$GITHUB_USERNAME
2. Go to your Grakn fork settings, and add `BAZEL_RBE_CREDENTIAL` environment variable with the RBE credential as the value: https://circleci.com/gh/$GITHUB_USERNAME/grakn/edit#env-vars

## Viewing Test Results From The RBE Build Status Page
When a test is executed, Bazel will print a URL which can be opened in the web browser to view the test status:
```
...
INFO: Streaming Build Event Protocol to https://source.cloud.google.com/results/invocations/dffc3106-e615-4120-b829-aa1f21e6e553
...
```
This contains roughly the same information as what's already there in CircleCI, but with a more Bazel-centric format.

# Why is this PR needed?
When a PR is submitted, the CI runs every tests even when the changes affect only a small portion of the codebase. Re-running everything has caused a waiting time of 30 - 45 mins on average no matter how small your PR is.

This PR closes https://github.com/graknlabs/grakn/issues/4574.

# What does the PR do?
Introduces Remote Build Execution where the test results are executed and cached across CI jobs. When a new PR is submitted and only changes a few parts of the code, only affected tests are re-ran. Tests which covers the code which is not changed won't be run.

Bazel is able to infer which tests needs to be re-run given a set of changes because it constructs a dependency graph from the `BUILD` definitions.

- Created a .bazelrc file which contains the Remote Build Execution configuration
- Invoke tests 
- Update bazel-common
- Depends on bazel-toolchain

# Does it break backwards compatibility?
Nein

# List of future improvements not on this PR
- Update //client-nodejs:test-integration to work with RBE #4773
- Update //client_python:test-integration to work with RBE #4774