Pattern negation pt 1.5 - unified negated pattern behaviour (#4812)
# Why is this PR needed?
Addresses #4234

Extends the set difference semantics to all patterns, i.e. we interpret executing queries with negated patterns in terms of performing set operations with negation corresponding to a set difference.

For a conjunctive query with negated patterns:

```
Q = A ∧ {...} ∧ ¬B ∧ ¬C ∧ {...}
```
Each answer x to query Q has to belong to the set:

``` 
{x : x ∈ A ∧ x !∈ B ∧ x !∈ C ∧ {...}}
``` 
or equivalently:

``` 
{x : x ∈ A x ∈ B^C ∧ x ∈ C^C ∧ {...}}
``` 
where upper C letter marks set complement.
 
As a result the answer set ans(Q) is equal to:
``` 
 ans(Q) = ans(A) \ [ ( ans(A) ∩ ans(B) ) ∪ ( ans(A) ∩ ans(C) ) ]
```
or equivalently
``` 
 ans(Q) = ans(A) ∩ ans(B^C) ∩ ans(C^C)
```

This aligns the semantics of a negated pattern with defining a negative literal with an atom corresponding to the body of the negation.
# What does the PR do?
Introduces `CompositeQuery` and `CompositeState` extending the ideas of `ReasonerQueryImpl` and `NegatedConjunctiveState` to all patterns.
# Does it break backwards compatibility?
No.
# List of future improvements not on this PR
Add support for negated disjunctions.