Refactor Graql query language library - part 13 (#4829)
# Why is this PR needed?

This PR is a continuation of PR #4646, and incremental work for issue #4456 where you can find the full description of the goal. In summary, we are refactoring Graql to extract the Graql language library from Grakn's implementation of Graql, and we start with simplifying the Graql codebase first.

# What does the PR do?

This PR focuses on removing any cyclic dependencies that prohibit us from extracting Graql language library from its implementation. In this case: removing the construction of `HasAttributeTypeProperty` internal data structure, namely the _"implicit statements"_. This is now moved to `HasAttributeTypeExecutor.matchFragments()`, which also allowed us to remove the `.innerImplicitStatement()` method from the `Statement` class.