Refactor Graql query language library - part 16 (#4892)
# Why is this PR needed?

This PR is a continuation of PR #4646, and incremental work for issue #4456 where you can find the full description of the goal. In summary, we are refactoring Graql to extract the Graql language library from Grakn's implementation of Graql, and we start with simplifying the Graql codebase first.

# What does the PR do?

1) Introduced Statement Builder classes to manage the diverse set of methods in Graql API to build Graql statements.
- Renamed `AttributeProperties` to `StatementAttributeBuilder`
- Moved builder methods from `Statements` into:
   * `builder.StatementTypeBuilder`
   * `builder.StatementInstancesBuilder`
   * `builder.StatementThingBuilder`
   * `builder.StatementRelationBuilder`
   * `builder.StatementAttributeBuilder`
- Split nested `StatementInstance` classes into its own files. (in prepartion for future improvements)

# Does it break backwards compatibility?

Nope.

# List of future improvements not on this PR

We will refine the Graql Statement data structure to be "strong typing" (issue #4899).