Fixed GraknClient (client-java) constructor (#4939)
## What is the goal of this PR?

GraknClient (in client-java) has several overloaded constructors. 2 of them are:
- `GraknClient(String host)`, which is the default one users should use, and
- `GraknClient(ManagedChannel channel)`, which allows us to override the channel by injection, useful for tests.

However, in user environments which requires explicit declaration of direct dependencies (even if they are already transitively loaded), e.g. in Bazel, this becomes an issue. The above constructor overloading makes `io.grpc.ManagedChannel` to be a direct dependency, which leads in a compile-time issue if the user is unaware. In most cases, this is unexpected, even though correct.

## What are the changes implemented in this PR?

The solution is to remove the second constructor `GraknClient(ManagedChannel channel)`, and provide a method to override the constructed channel. Although this is not so ideal, it's the lesser of all evils.