Return Answer Explanation in client-java (#4950)
## What is the goal of this PR?

In the process of completing PR #4942 and #4926, we needed to implement Answer Explanations for inferred answers of Graql queries in client-java (issue #3155), which is now possible because of PR #4948.

## What are the changes implemented in this PR?

Fixes #3155: Return `Answer` `Explanation` for inferred answers Graql queries in client-java
Fixes #4197: Change `Explanation` class to store the query as `Pattern` rather than `ReasonerQuery`. PR #4948 already replaces the `ReasonerQuery`s with `String`s, but we replace it again with `Pattern` as it would be more robust.