Implemented case-insensitive 'contains' predicate (#4988)
## What is the goal of this PR?

As reported by issue #4831, the `contains` predicate at the moment is not case-insensitive, which is not very useful when searching for text flexibly.

## What are the changes implemented in this PR?

Fixes #4831: We've implemented a `containsIgnoreCase(String str1, String str2)` method in `ValueExecutor` that compares two strings regardless of their case. This implementation is a more efficient implementation, which is important given that this method will be called at every iteration of a stream consumption of the predicate. I.e. it can get really expensive really quickly if not implemented efficiently. 

This implementation uses `str1.regionMatches(str2)` from a specific position in `str1`, while the other two alternatives are:
- `(str1, str1) -> str1.toLowerCase().contains(str2.toLowerCase())`
- several variations of using `java.util.regex.Pattern`