Enforcing unique usage of keys (#4995)
## What is the goal of this PR?

At the moment, attributes that are used as keys are not guaranteed to be used at most once within the set of all instances of a given type (issue #4928).

## What are the changes implemented in this PR?

Fixes #4928: in `ValidateGlobalRules.validateInstancePlaysAllRequiredRoles(Thing)`, we already validate that an instance of a given type with a key, needs to have exactly one key. However, we forgot to verify that the key is used not more than once in the set of all instances of that type. We've implemented this check and added some tests.