Order deletes - remove relations first (#5109)
## What is the goal of this PR?
Fix issue when deleting multiple concepts in one query, where some of them are concepts that are role players and others are the relations relating these RPs. When we delete concepts playing roles, if the relations they are in have 0 role players, we delete them too. This means that sometimes when deleting RPs and relations, the relation is already gone by the time we want to delete it explicitly.

Solution: always delete relations before other concepts. This may still hit an edge case and cause an exception when deleting relations that are role players in other relations. To solve this, a try/catch was added and the event is logged.

## What are the changes implemented in this PR?
* Add an ordering to delete relations first
* Try/Catch around deleting concepts already deleted - should only occur when deleting RPs that are relations.