Caches cleanup part 2 (#5113)
Removed:

Cacheable class, which was a helper used to copy a cached value (that was valid within the current transaction) to a valueGlobal container which was used to store values that could be reused by other transactions in the same session. We enforce all concepts to be transaction bound and so should be the caches used by concepts.

Removed difference between sessionCache and transactionCache given that we don't cache any concept in sessions anymore.

Removed cache invalidation when closing a transaction because it should not be necessary anymore to do so, again because concepts and their caches should only be used within the current transaction.