Synchronize Graql - Extended Label grammar and renamed `asUserDefined` (#5128)
## What is the goal of this PR?
Synchronize with a new version of Graql, which supports extended character sets for labels and implicit labels, and renames `asUserDefined` to `asReturnedVar` no `Variable` objects.
 
## What are the changes implemented in this PR?
Synchronize with changes in the Graql repository. Code had to be updated in various places to rename `var.asUserDefined()` and `var.isUserDefinedName()` to `var.asReturnedVar()` and `var.isReturned()`.