Close session if client dies (#5158)
## What is the goal of this PR?

With this PR we cover the use case of a client terminating connection to the server abruptly.

It might happen that a user forces the termination of the client application, in that case the `onCompleted()` method in `TransactionListener` and the `close()` method in `SessionService` will not be invoked, this will result in an open session that will remain forever in memory.

Forcing a cleanup of session in `onError()` in `TransactionListener` we make sure not to leave any unused open session in memory.

The only way to end up with an unused open session in memory would be to call `onCompleted()` in `TransactionListener` and then terminating abruptly the client-application, but this is way less likely to happen. To cover this case we would need to reimplement `SessionService` to be a full duplex stream as it currently is `transaction()` service.

## What are the changes implemented in this PR?

Store associated `sessionId` in `TransactionListener` and close the session in `onError()`.