Make sure PID stored in pid files is actually bound to a Grakn process (#5192)
## What is the goal of this PR?

We now make sure that any PID stored inside temp files `grakn-server.pid` and `grakn-storage.pid` is currently running AND it is also bound to a Grakn/GraknStorage process.

This is because it might happen that Grakn does not terminate gracefully (e.g. restart machine), in which case the `.pid` files won't be deleted. 
Before this PR Grakn would try to start reading PIDs from those files and in case the OS reassigned the PIDs to new processes, Grakn would be tricked into believing that other Grakn processes are already running. 

## What are the changes implemented in this PR?

- on top of checking if PID stored in file is running make sure that the `Command` that started the associated process contains the string `core.grakn.server.Grakn` / `core.grakn.server.GraknServer`

- remove unused functions and variables