Statistics - track total Thing count (#5212)
## What is the goal of this PR?
Maintain a count of the total number of Things committed to the graph (this includes implicit relations). This should be equal to the number of concepts counted by an aggregate count in a new transaction. This means that retrieving the total number of things persisted no longer requires traversing all the `subs()` of `Thing` and adding the counts of each of these, as this can be an expensive operation if the schema is large.

## What are the changes implemented in this PR?
* On commit, we sum the total of the concrete instance types changed and add them to the label `Thing`, which is persisted into the meta `Thing` concept just like any other schema concept would be handled.
* Tests updated to check the count of `Thing`
* Replace hardcoded upper bound in `InstanceNode` to use the `Thing` count from statistics (it was too expensive to do all the `subs()` before --  see above)