Add read through attributesMap in a thread safe way (#5257)
## What is the goal of this PR?
Fix attribute deduplicator from dieing when sessions are cleaned up and the attribute map is empty. This is achieved using a synchronized read-through to Janus when the attribute map does not contain the required attribute yet.

## What are the changes implemented in this PR?
A small read-through using `.compute()` on the synchronized attributes hash map, that reads the ID of the desired attribtue from Janus.