Make sure we do not discard rules if type is absent in negation block (#5256)
## What is the goal of this PR?
In the `RuleCache`, as rules are being assessed for matchability, we scan the types for possible absence of instances which allows us to discard rules that yield no answers. However! If a rule has negation blocks in it, absence of instances yields results! Therefore we correct the pruning behaviour to take into account rules with negation.

## What are the changes implemented in this PR?
Scanning types for absence of instances when determining rule matchability is done only for positive types - ones that are present in non-negated blocks of the `when` part.