Refactor attribute atom materialisation pt II (#5246)
## What is the goal of this PR?
When materialising attributes we want to follow a similar procedure to materialising relations. This means we try to use graql for checking for answer existence. In that way we can cache the checks and reuse them later.

## What are the changes implemented in this PR?
Employs a similar materialisation procedure for attributes as we currently have for relations - we employ the put behaviour by having a cache lookup before a potential insert.