Limited cache (#5278)
## What is the goal of this PR?

Change attributesMap to attributesCache and make it a Cache with limited size instead of a ConcurrentHashMap which could potentially grow out of memory.

## What are the changes implemented in this PR?

The new cache is limited to `10000` attributes and each key in the cache expires after 2 minutes.
We rely on the fact that after 2 minutes JanusGraph should correctly report as committed any attribute that has been in the graph for 2 minutes. This still require further investigation on Janus guarantees after committing an attribute with `INDEX` set. 

- change name `attributesMap` to `attributesCache`
- change implementation from `ConcurrentHashMap` to  Guava `Cache`